Trac #32867: remove the MAXIMA sage_conf variable.
Until now, the MAXIMA variable held the name of the main maxima
executable. But there really is no reason to use anything other
than "maxima -l ecl", which works in all cases. This commit
replaces the variable with the hard-coded value. This simplifies
the use of maxima from the system because we no longer need to
set sage_conf.MAXIMA to any particular value during ./configure.