Trac #32981: use CPU time for --warn-long.
Our --warn-long flag that is used during testing currently takes "wall
time" as its argument. This is somewhat nice from a user perspective,
since a normal user is quite likely to know what what "wall time" is,
but it's also a bit nonsensical given the purpose of the flag. With
the flag being designed to warn developers about tests that take too
long, whether or not it triggers should not depend on the developer's
machine or what else he's doing at the time. In fact, when no value
for --wall-time is given, we already try to use a heuristic that turns
the default into a CPU-time equivalent.

This commit changes the flag to use CPU time instead, obviating the
need for the heuristic. We also update the relevant documentation to
clarify that tests should run in a certain amount of CPU time -- not
in some given amount of wall time.