Trac #33176: update API usage in sage.libs.gap.element.
The C_NEW_STRING() macro in GAP's stringobj.h is deprecated (according
to the source), and should be replaced with MakeStringWithLen(). So
that's what we do. The conversion is straightforward and happens to
fix the Cython warnings,

  warning: sage/libs/gap/element.pyx:267:21: local variable 'result'
  referenced before assignment
  warning: sage/libs/gap/element.pyx:268:15: local variable 'result'
  referenced before assignment

because now the return value of MakeStringWithLen() is assigned to
the "result" variable.