Trac #33134: remove a failing doctest in sage.rings.polynomial.hilbert.
A test for hilbert_numerator() in this module (a) can fail with older
versions of singular, and (b) appears buggy with new versions of
singular. Here we remove the test to avoid the inevitable failures. A
follow-up ticket has been opened (Trac 33178) to address the
underlying problem, which may take more time.